# PERF-#4705: Improve perf of arithmetic operations between `Series` objects with shared `.index`