```
import modin.config as cfg
cfg.BenchmarkMode.put(True)

import ray
ray.init()
import modin.pandas as pd
import numpy as np

arr = np.random.randn(100_000, 2)
df = pd.DataFrame(arr)
df.index = pd.MultiIndex.from_product([list('abcdefghij'), np.arange(10_000)])
df = pd.concat([df]*10)

In [4]: %timeit df[0] + df[1]
337 ms ± 6.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
264 ms ± 6.93 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR

In [5]: df = pd.concat([df]*10)

In [6]: %timeit df[0] + df[1]
3.29 s ± 132 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
2.63 s ± 334 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR
```