@pyrito 2022-07-20
Thanks for the PR @jbrockmendel !

Could you try running your performance measurements again with benchmark mode enabled?
```python
import modin.config as cfg

# Enable benchmark mode
cfg.BenchmarkMode.put(True)
```