@jbrockmendel 2022-07-21
With BenchmarkMode enabled I get 

```
In [4]: %timeit df[0] + df[1]
337 ms ± 6.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
264 ms ± 6.93 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR
```