@pyrito 2022-07-21
@jbrockmendel could you try running the benchmarks for larger datasets and let us know how the performance looks there?

Also, it looks like the rest of the CI runs aren't running because your commit isn't formatted properly. You'll probably have to create a Github issue first and then link this PR to that. You can check this link out for more details: https://modin.readthedocs.io/en/stable/development/contributing.html