@jbrockmendel 2022-07-22
Updated OP with results on 10x larger DataFrame.